requirements.txt
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0
apscheduler>=3.8.0 
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-cors==3.0.10
flask-jwt-extended==4.3.1
gunicorn==20.1.0
sqlalchemy==1.4.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
fastapi>=0.109.0
pydantic>=2.6.0
starlette>=0.36.0
uvicorn>=0.27.0 